requirements.txt
mysql-connector-python~=8.0
